+++ app/workers/bq_analyze_tweeters_by_topic.py
@@ -8,6 +8,7 @@ from scipy.stats import kstest, ks_2samp
 from dotenv import load_dotenv
 
 from app.bq_service import BigQueryService
+from app.workers import to_dt, to_ts
 
 load_dotenv()
 np.random.seed(2020)
@@ -17,25 +18,6 @@ Y_TOPIC = os.getenv("Y_TOPIC", default="#ImpeachAndRemove")
 
 X_LIMIT = int(os.getenv("X_LIMIT", default="1500")) # sample size
 Y_LIMIT = int(os.getenv("Y_LIMIT", default="1500")) # sample size
-
-def to_ts(my_dt):
-    """
-    Converts datetime object to timestamp (seconds since epoch).
-
-    Param: dt (datetime) like ... datetime.datetime(2016, 7, 23, 10, 38, 35, 636364)
-
-    """
-    return int(my_dt.timestamp())
-
-def to_dt(ts):
-    """
-    Converts timestamp (seconds since epoch) to datetime object.
-
-    Param: dti (int) like ... 1469270315.6363637
-
-    Returns:
-    """
-    return datetime.utcfromtimestamp(ts)
 
 if __name__ == "__main__":
